backend/ChamCong/face_recognition_system/requirements.txt
flask
flask-sqlalchemy
opencv-python
face-recognition
numpy
Pillow
python-dateutil
mysql-connector-python
torch
torchvision
easydict
tqdm
tensorboardX